pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pickingsystem"
version = "0.1.0"
description = "Picking System desktop UI with local API"
readme = "README.md"
authors = [{name = "Auto Generated"}]
requires-python = ">=3.11"
dependencies = [
  "PySide6",
  "fastapi",
  "uvicorn[standard]",
  "pandas",
  "openpyxl",
  "jinja2",
  "requests",
  "pydantic",
  "pydantic-settings",
  "PyYAML",
  "watchdog",
  "rich",
  "playwright",
  "pywin32",
  "qrcode[pil]"
]

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "pytest",
  "ruff",
  "types-PyYAML",
  "types-requests"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
mypy_path = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
